# farin-main/docs/swagger.yaml
definitions:
...
  farin_app_api_response.Response-farin_domain_dto_RingListResponse:
    properties:
      data:
        $ref: '#/definitions/farin_domain_dto.RingListResponse'
      message:
        type: string
      status:
        type: integer
    type: object
  farin_app_api_response.Response-farin_domain_dto_RingResponse:
    properties:
      data:
        $ref: '#/definitions/farin_domain_dto.RingResponse'
      message:
        type: string
      status:
        type: integer
    type: object
...
  farin_app_api_response.Response-swagger_EmptyObject:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyObject'
      message:
        type: string
      status:
        type: integer
    type: object
...
  farin_domain_dto.RingListResponse:
    properties:
      rings:
        items:
          $ref: '#/definitions/farin_domain_dto.RingResponse'
        type: array
      total:
        type: integer
    type: object
  farin_domain_dto.RingRequest:
    properties:
      geom:
        description: Assuming geometry is stored as a string (e.g., WKT format)
        type: string
      id:
        type: integer
      length:
        minimum: 0
        type: number
      ringCode:
        maxLength: 100
        minLength: 1
        type: string
      ringName:
        maxLength: 120
        minLength: 1
        type: string
    required:
    - geom
    - id
    - length
    - ringCode
    - ringName
    type: object
  farin_domain_dto.RingResponse:
    properties:
      createdAt:
        type: integer
      geom:
        type: string
      id:
        type: integer
      length:
        type: number
      ringCode:
        type: string
      ringName:
        type: string
      updatedAt:
        type: integer
    type: object
...
  swagger.EmptyObject:
    type: object
info:
  contact: {}
paths:
...
  /rings:
    get:
      consumes:
      - application/json
      description: Retrieve a list of rings with optional filters and pagination
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - default: created_at
        description: Field to sort by
        in: query
        name: sortField
        type: string
      - default: asc
        description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_RingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: List rings
      tags:
      - rings
    post:
      consumes:
      - application/json
      description: Create a new ring by providing ring details
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ring Data
        in: body
        name: ring
        required: true
        schema:
          $ref: '#/definitions/farin_domain_dto.RingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_RingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Create a new ring
      tags:
      - rings
  /rings/{id}:
    delete:
      description: Delete a ring by ID
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ring ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Delete a ring
      tags:
      - rings
    get:
      description: Retrieve ring details by ID
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ring ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_RingResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Get ring details
      tags:
      - rings
    put:
      consumes:
      - application/json
      description: Update an existing ring's details
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated Ring Data
        in: body
        name: ring
        required: true
        schema:
          $ref: '#/definitions/farin_domain_dto.RingRequest'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_RingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Update ring details
      tags:
      - rings
...
swagger: "2.0"